add comment

+++ src/main/java/cn/zh/community/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
      * @return
      */
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -51,7 +51,7 @@ public class PublishController {
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
             //接收publish.html中<input type="hidden" name="id" th:value="${id}">的id值
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model) {
